chore: move hook into network selector folder
